+++ src/simulation/rkf_parameters.cuh
@@ -6,17 +6,17 @@
 
 struct RKFParameters
 {
-    Float initial_time_step = 1;
-    Float min_time_step = 1e-6;
-    Float abs_tol = 1e-9;
-    Float rel_tol = 1e-9;
-    Float scale_state = 1;
-    Float scale_dstate = 0;
+    double initial_time_step = 1;
+    double min_time_step = 1e-6;
+    double abs_tol = 1e-9;
+    double rel_tol = 1e-9;
+    double scale_state = 1;
+    double scale_dstate = 0;
     std::size_t max_steps = 1e6;
 
-    static constexpr Float dt_safety = .9;
-    static constexpr Float min_dt_scale = .2;
-    static constexpr Float max_dt_scale = 5;
+    static constexpr double dt_safety = .9;
+    static constexpr double min_dt_scale = .2;
+    static constexpr double max_dt_scale = 5;
 
     static RKFParameters from_json(const nlohmann::json &exec_json);
 };
